module-6-7-8-9/services/preview/requirements.txt
Django>=5.0
djangorestframework
httpx
beautifulsoup4
psycopg2-binary
django-environ>=0.11
drf-spectacular>=0.27.0
tenacity>=8.2.0
pybreaker>=1.2.0
